requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers>=3.0.0
llama-cpp-python==0.2.28
pypdf==3.17.4
python-docx==1.1.0
requests==2.31.0
pyyaml==6.0.1
huggingface-hub==0.20.3
